pyproject.toml
[project]
name = "clock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gpiozero>=2.0.1",
    "lgpio>=0.2.2.0",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "spidev>=3.8",
]
